// src/@types/PreVenda.d.ts
export interface iItemPreVenda {
  CodigoProduto: string;
  Qtd: number;
  Valor: number;
  SubTotal: number;
  Frete: number;
  Desconto: number;
  Total: number;
  Tabela: string;
}
export interface iPreVenda {
  DataPedido: string;
  CodigoCliente: number;
  CodigoVendedor1: number;
  CodigoVendedor2?: number;
  SubTotal: number;
  ValorFrete?: number;
  Desconto?: number;
  Total: number;
  ObsPedido1: string;
  ObsPedido2?: string;
  ObsNotaFiscal: string;
  CodigoCondicaoPagamento: number;
  PedidoEcommerce?: string;
  ModeloNota: string;
  NumeroOrdemCompraCliente: string;
  Entrega: 'S' | 'N';
  TipoEntrega?: 'VEM BUSCAR' | 'CARRO';
  Origem?: string;
  Itens: iItemPreVenda[];
}

export interface iParcelasPgto {
  DIAS: number;
  VENCIMENTO: string;
  VALOR: number;
}

export interface iCondicaoPgto {
  ID: number;
  NOME: string;
  PARCELAS: number;
  VALOR_PARCELA: number;
  VALOR_MINIMO: number;
  PM: number;
  PZ01: number;
  PZ02: number;
  PZ03: number;
  PZ04: number;
  PZ05: number;
  PZ06: number;
  PZ07: number;
  PZ08: number;
  PZ09: number;
  PZ10: number;
  TIPO: string;
  DESTACAR_DESCONTO: string;
  FORMA: string;
}

export interface iFormaPgto {
  CARTAO: string;
}